Dovecot based version of the mailserver. Courier and Cyrus Sasl have been removed and substituted with Dovecot which now handle authentication for Postfix, Imap and Pop3, with support for SSL. This allow the use of several encryption schemes for the password as well as a single user db. OpenDKIM keys can now be provided at the startup and will be used instead of generating new ones (so that you don't have to change your DNS configuration).
This version builds correctly on Docker but no integration tests have been
reworked to accommodate Dovecot instead of Courier and Cyrus Sasl. As such
at present no automatic tests can be executed.

+++ start-mailserver.sh
@@ -11,8 +11,20 @@ if [ -f /tmp/postfix/accounts.cf ]; then
 
   # Checking that /tmp/postfix/accounts.cf ends with a newline
   sed -i -e '$a\' /tmp/postfix/accounts.cf
+  # Configuring Dovecot
+  echo -n > /etc/dovecot/userdb
+  chown dovecot:dovecot /etc/dovecot/userdb
+  chmod 640 /etc/dovecot/userdb
+  cp -a /usr/share/dovecot/protocols.d /etc/dovecot/
+  # Disable pop3 (it will be eventually enabled later in the script, if requested)
+  mv /etc/dovecot/protocols.d/pop3d.protocol /etc/dovecot/protocols.d/pop3d.protocol.disab
+  sed -i -e 's/#ssl = yes/ssl = yes/g' /etc/dovecot/conf.d/10-master.conf
+  sed -i -e 's/#port = 993/port = 993/g' /etc/dovecot/conf.d/10-master.conf
+  sed -i -e 's/#port = 995/port = 995/g' /etc/dovecot/conf.d/10-master.conf
+  sed -i -e 's/#ssl = yes/ssl = required/g' /etc/dovecot/conf.d/10-ssl.conf
 
   # Creating users
+  # pass is encrypted
   while IFS=$'|' read login pass
   do
     # Setting variables for better readability
@@ -21,22 +33,22 @@ if [ -f /tmp/postfix/accounts.cf ]; then
     # Let's go!
     echo "user '${user}' for domain '${domain}' with password '********'"
     echo "${login} ${domain}/${user}/" >> /etc/postfix/vmailbox
-    /usr/sbin/userdb ${login} set uid=5000 gid=5000 home=/var/mail/${domain}/${user} mail=/var/mail/${domain}/${user}
-    echo "${pass}" | userdbpw -md5 | userdb ${login} set systempw
-    echo "${pass}" | saslpasswd2 -p -c -u ${domain} ${login}
+    # user database for dovecot has the following format:
+    # user:password:uid:gid:(gecos):home:(shell):extra_fields
+    # Example : ${login}:${pass}:5000:5000::/var/mail/${domain}/${user}::userdb_mail=maildir:/var/mail/${domain}/${user}
+    echo "${login}:${pass}:5000:5000::/var/mail/${domain}/${user}::" >> /etc/dovecot/userdb
     mkdir -p /var/mail/${domain}
     if [ ! -d "/var/mail/${domain}/${user}" ]; then
-      maildirmake "/var/mail/${domain}/${user}"
-      maildirmake "/var/mail/${domain}/${user}/.Sent"
-      maildirmake "/var/mail/${domain}/${user}/.Trash"
-      maildirmake "/var/mail/${domain}/${user}/.Drafts"
-      echo -e "INBOX\nINBOX.Sent\nINBOX.Trash\nInbox.Drafts" >> "/var/mail/${domain}/${user}/courierimapsubscribed"
+      maildirmake.dovecot "/var/mail/${domain}/${user}"
+      maildirmake.dovecot "/var/mail/${domain}/${user}/.Sent"
+      maildirmake.dovecot "/var/mail/${domain}/${user}/.Trash"
+      maildirmake.dovecot "/var/mail/${domain}/${user}/.Drafts"
+      echo -e "INBOX\nSent\nTrash\nDrafts" >> "/var/mail/${domain}/${user}/subscriptions"
       touch "/var/mail/${domain}/${user}/.Sent/maildirfolder"
 
     fi
     echo ${domain} >> /tmp/vhost.tmp
   done < /tmp/postfix/accounts.cf
-  makeuserdb
 else
   echo "==> Warning: '/tmp/postfix/accounts.cf' is not provided. No mail account created."
 fi
@@ -63,39 +75,46 @@ touch /etc/postfix/vmailbox && postmap /etc/postfix/vmailbox
 touch /etc/postfix/virtual && postmap /etc/postfix/virtual
 
 # DKIM
-grep -vE '^(\s*$|#)' /etc/postfix/vhost | while read domainname; do
-  mkdir -p /etc/opendkim/keys/$domainname
-  if [ ! -f "/etc/opendkim/keys/$domainname/mail.private" ]; then
-    echo "Creating DKIM private key /etc/opendkim/keys/$domainname/mail.private"
-    pushd /etc/opendkim/keys/$domainname
-    opendkim-genkey --subdomains --domain=$domainname --selector=mail
-    popd
-    echo ""
-    echo "DKIM PUBLIC KEY ################################################################"
-    cat /etc/opendkim/keys/$domainname/mail.txt
-    echo "################################################################################"
-  fi
-  # Write to KeyTable if necessary
-  keytableentry="mail._domainkey.$domainname $domainname:mail:/etc/opendkim/keys/$domainname/mail.private"
-  if [ ! -f "/etc/opendkim/KeyTable" ]; then
-    echo "Creating DKIM KeyTable"
-    echo "mail._domainkey.$domainname $domainname:mail:/etc/opendkim/keys/$domainname/mail.private" > /etc/opendkim/KeyTable
-  else
-    if ! grep -q "$keytableentry" "/etc/opendkim/KeyTable" ; then
-      echo $keytableentry >> /etc/opendkim/KeyTable
+# Check if keys are already available
+if [ -e "/tmp/postfix/opendkim/KeyTable" ]; then
+  mkdir -p /etc/opendkim
+  cp -a /tmp/postfix/opendkim/* /etc/opendkim/
+  echo "DKIM keys added for : `ls -C /etc/opendkim/keys/`"
+else 
+  grep -vE '^(\s*$|#)' /etc/postfix/vhost | while read domainname; do
+    mkdir -p /etc/opendkim/keys/$domainname
+    if [ ! -f "/etc/opendkim/keys/$domainname/mail.private" ]; then
+      echo "Creating DKIM private key /etc/opendkim/keys/$domainname/mail.private"
+      pushd /etc/opendkim/keys/$domainname
+      opendkim-genkey --subdomains --domain=$domainname --selector=mail
+      popd
+      echo ""
+      echo "DKIM PUBLIC KEY ################################################################"
+      cat /etc/opendkim/keys/$domainname/mail.txt
+      echo "################################################################################"
     fi
-  fi
-  # Write to SigningTable if necessary
-  signingtableentry="*@$domainname mail._domainkey.$domainname"
-  if [ ! -f "/etc/opendkim/SigningTable" ]; then
-    echo "Creating DKIM SigningTable"
-    echo "*@$domainname mail._domainkey.$domainname" > /etc/opendkim/SigningTable
-  else
-    if ! grep -q "$signingtableentry" "/etc/opendkim/SigningTable" ; then
-      echo $signingtableentry >> /etc/opendkim/SigningTable
+    # Write to KeyTable if necessary
+    keytableentry="mail._domainkey.$domainname $domainname:mail:/etc/opendkim/keys/$domainname/mail.private"
+    if [ ! -f "/etc/opendkim/KeyTable" ]; then
+      echo "Creating DKIM KeyTable"
+      echo "mail._domainkey.$domainname $domainname:mail:/etc/opendkim/keys/$domainname/mail.private" > /etc/opendkim/KeyTable
+    else
+      if ! grep -q "$keytableentry" "/etc/opendkim/KeyTable" ; then
+        echo $keytableentry >> /etc/opendkim/KeyTable
+      fi
     fi
-  fi
-done
+    # Write to SigningTable if necessary
+    signingtableentry="*@$domainname mail._domainkey.$domainname"
+    if [ ! -f "/etc/opendkim/SigningTable" ]; then
+      echo "Creating DKIM SigningTable"
+      echo "*@$domainname mail._domainkey.$domainname" > /etc/opendkim/SigningTable
+    else
+      if ! grep -q "$signingtableentry" "/etc/opendkim/SigningTable" ; then
+        echo $signingtableentry >> /etc/opendkim/SigningTable
+      fi
+    fi
+  done
+fi
 
 echo "Changing permissions on /etc/opendkim"
 # chown entire directory
@@ -104,7 +123,7 @@ chown -R opendkim:opendkim /etc/opendkim/
 chmod -R 0700 /etc/opendkim/keys/
 
 # DMARC
-# if ther is no AuthservID create it
+# if there is no AuthservID create it
 if [ `cat /etc/opendmarc.conf | grep -w AuthservID | wc -l` -eq 0 ]; then
   echo "AuthservID $(hostname)" >> /etc/opendmarc.conf
 fi
@@ -129,15 +148,9 @@ case $DMS_SSL in
       sed -i -r 's/smtpd_tls_cert_file=\/etc\/ssl\/certs\/ssl-cert-snakeoil.pem/smtpd_tls_cert_file=\/etc\/letsencrypt\/live\/'$(hostname)'\/fullchain.pem/g' /etc/postfix/main.cf
       sed -i -r 's/smtpd_tls_key_file=\/etc\/ssl\/private\/ssl-cert-snakeoil.key/smtpd_tls_key_file=\/etc\/letsencrypt\/live\/'$(hostname)'\/privkey.pem/g' /etc/postfix/main.cf
 
-      # Courier configuration
-      cat "/etc/letsencrypt/live/$(hostname)/cert.pem" "/etc/letsencrypt/live/$(hostname)/chain.pem" "/etc/letsencrypt/live/$(hostname)/privkey.pem" > "/etc/letsencrypt/live/$(hostname)/combined.pem"
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/imapd.pem/TLS_CERTFILE=\/etc\/letsencrypt\/live\/'$(hostname)'\/combined.pem/g' /etc/courier/imapd-ssl
-
-      # POP3 courier configuration
-      sed -i -r 's/POP3_TLS_REQUIRED=0/POP3_TLS_REQUIRED=1/g' /etc/courier/pop3d-ssl
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/pop3d.pem/TLS_CERTFILE=\/etc\/letsencrypt\/live\/'$(hostname)'\/combined.pem/g' /etc/courier/pop3d-ssl
-      # needed to support gmail
-      sed -i -r 's/TLS_TRUSTCERTS=\/etc\/ssl\/certs/TLS_TRUSTCERTS=\/etc\/letsencrypt\/live\/'$(hostname)'\/fullchain.pem/g' /etc/courier/pop3d-ssl
+      # Dovecot configuration
+      sed -i -e 's/ssl_cert = <\/etc\/dovecot\/dovecot\.pem/ssl_cert = <\/etc\/letsencrypt\/live\/'$(hostname)'\/fullchain\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
+      sed -i -e 's/ssl_key = <\/etc\/dovecot\/private\/dovecot\.pem/ssl_key = <\/etc\/letsencrypt\/live\/'$(hostname)'\/privkey\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
 
       echo "SSL configured with letsencrypt certificates"
 
@@ -154,12 +167,9 @@ case $DMS_SSL in
       sed -i -r 's/smtpd_tls_cert_file=\/etc\/ssl\/certs\/ssl-cert-snakeoil.pem/smtpd_tls_cert_file=\/etc\/postfix\/ssl\/'$(hostname)'-full.pem/g' /etc/postfix/main.cf
       sed -i -r 's/smtpd_tls_key_file=\/etc\/ssl\/private\/ssl-cert-snakeoil.key/smtpd_tls_key_file=\/etc\/postfix\/ssl\/'$(hostname)'-full.pem/g' /etc/postfix/main.cf
 
-      # Courier configuration
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/imapd.pem/TLS_CERTFILE=\/etc\/postfix\/ssl\/'$(hostname)'-full.pem/g' /etc/courier/imapd-ssl
-
-      # POP3 courier configuration
-      sed -i -r 's/POP3_TLS_REQUIRED=0/POP3_TLS_REQUIRED=1/g' /etc/courier/pop3d-ssl
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/pop3d.pem/TLS_CERTFILE=\/etc\/postfix\/ssl\/'$(hostname)'-full.pem/g' /etc/courier/pop3d-ssl
+      # Dovecot configuration
+      sed -i -e 's/ssl_cert = <\/etc\/dovecot\/dovecot\.pem/ssl_cert = <\/etc\/postfix\/ssl\/'$(hostname)'-full\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
+      sed -i -e 's/ssl_key = <\/etc\/dovecot\/private\/dovecot\.pem/ssl_key = <\/etc\/postfix\/ssl\/'$(hostname)'-full\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
 
       echo "SSL configured with CA signed/custom certificates"
 
@@ -176,6 +186,8 @@ case $DMS_SSL in
       mkdir -p /etc/postfix/ssl
       cp "/tmp/postfix/ssl/$(hostname)-cert.pem" /etc/postfix/ssl
       cp "/tmp/postfix/ssl/$(hostname)-key.pem" /etc/postfix/ssl
+      # Force permission on key file
+      chmod 600 /etc/postfix/ssl/$(hostname)-key.pem
       cp "/tmp/postfix/ssl/$(hostname)-combined.pem" /etc/postfix/ssl
       cp /tmp/postfix/ssl/demoCA/cacert.pem /etc/postfix/ssl
 
@@ -186,12 +198,9 @@ case $DMS_SSL in
       sed -i -r 's/#smtp_tls_CAfile=/smtp_tls_CAfile=\/etc\/postfix\/ssl\/cacert.pem/g' /etc/postfix/main.cf
       ln -s /etc/postfix/ssl/cacert.pem "/etc/ssl/certs/cacert-$(hostname).pem"
 
-      # Courier configuration
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/imapd.pem/TLS_CERTFILE=\/etc\/postfix\/ssl\/'$(hostname)'-combined.pem/g' /etc/courier/imapd-ssl
-
-      # POP3 courier configuration
-      sed -i -r 's/POP3_TLS_REQUIRED=0/POP3_TLS_REQUIRED=1/g' /etc/courier/pop3d-ssl
-      sed -i -r 's/TLS_CERTFILE=\/etc\/courier\/pop3d.pem/TLS_CERTFILE=\/etc\/postfix\/ssl\/'$(hostname)'-combined.pem/g' /etc/courier/pop3d-ssl
+      # Dovecot configuration
+      sed -i -e 's/ssl_cert = <\/etc\/dovecot\/dovecot\.pem/ssl_cert = <\/etc\/postfix\/ssl\/'$(hostname)'-combined\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
+      sed -i -e 's/ssl_key = <\/etc\/dovecot\/private\/dovecot\.pem/ssl_key = <\/etc\/postfix\/ssl\/'$(hostname)'-key\.pem/g' /etc/dovecot/conf.d/10-ssl.conf
 
       echo "SSL configured with self-signed/custom certificates"
 
@@ -223,7 +232,6 @@ fi
 echo "Fixing permissions"
 chown -R 5000:5000 /var/mail
 mkdir -p /var/log/clamav && chown -R clamav:root /var/log/clamav
-chown postfix.sasl /etc/sasldb2
 
 echo "Creating /etc/mailname"
 echo $(hostname -d) > /etc/mailname
@@ -254,19 +262,18 @@ sed -i -r "/^#?compress/c\compress\ncopytruncate" /etc/logrotate.conf
 echo "Starting daemons"
 cron
 /etc/init.d/rsyslog start
-/etc/init.d/saslauthd start
+##/etc/init.d/saslauthd start
 
 if [ "$SMTP_ONLY" != 1 ]; then
-
-/etc/init.d/courier-authdaemon start
-/etc/init.d/courier-imap start
-/etc/init.d/courier-imap-ssl start
-
+  # Here we are starting sasl and imap, not pop3 because it's disabled by default
+  echo " * Starting dovecot services"
+  /usr/sbin/dovecot -F -c /etc/dovecot/dovecot.conf &
 fi
+
 if [ "$ENABLE_POP3" = 1 -a "$SMTP_ONLY" != 1 ]; then
   echo "Starting POP3 services"
-  /etc/init.d/courier-pop start
-  /etc/init.d/courier-pop-ssl start
+  mv /etc/dovecot/protocols.d/pop3d.protocol.disab /etc/dovecot/protocols.d/pop3d.protocol
+  /usr/sbin/dovecot reload
 fi
 
 /etc/init.d/spamassassin start
@@ -281,8 +288,8 @@ if [ "$ENABLE_FAIL2BAN" = 1 ]; then
   /etc/init.d/fail2ban start
 fi
 
-echo "Listing SASL users"
-sasldblistusers2
+echo "Listing users"
+/usr/sbin/dovecot user '*'
 
 echo "Starting..."
 tail -f /var/log/mail.log
